updated some css settings
