+++ its/src/Component/ManagementConsole.css
@@ -1,0 +1,6 @@
+.ManagementConsole {
+    margin-top: 100px;
+    margin-bottom: 100px;
+    margin-right: 150px;
+    margin-left: 80px;
+}
